Criação endpoint Motos e ajustes gerais

+++ source/Components/Desafio-BackEnd.Domain.Core/Queries/QueryResult.cs
@@ -17,7 +17,7 @@ namespace Desafio_BackEnd.Domain.Core.Queries
             TotalRegistros = Registros.Count;
         }
 
-        public QueryResult(int paginaAtual, int registrosPorPagina, int totalRegistros, ICollection<T> registros)
+        public QueryResult(int paginaAtual, int registrosPorPagina, long totalRegistros, ICollection<T> registros)
         {
             PaginaAtual = paginaAtual;
             RegistrosPorPagina = registrosPorPagina;
